Change zoom property type to "float"
Done to reflect eclipsesource/tabris-maps@f536a03.

+++ www/mapWidget.js
@@ -6,7 +6,7 @@ tabris.registerWidget("Map", {
   _properties: {
     liteMode: "boolean",
     latLng: "array",
-    zoom: "integer",
+    zoom: "float",
     mapType: {type: ["choice", ["none", "hybrid", "normal", "satellite", "terrain"]]}
   }
 });
